pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daps-env-v2"
version = "2.0.0"
description = "Vanguard Digital Asset Protection System (DAPS) for Meta x PyTorch OpenEnv Hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    "openai",
    "requests",
    "torch",
    "torchvision",
    "faiss-cpu",
    "imagehash",
    "Pillow",
    "sqlalchemy",
    "numpy",
    "scipy",
]

[project.scripts]
server = "openenv.core.cli:server"

[project.entry-points."openenv.environments"]
daps-env-v2 = "server.app:app"

[tool.setuptools]
packages = ["server", "server.db", "server.core", "server.scripts"]
py-modules = ["inference"]
